+++ documentation/src/pages/about/index.tsx
@@ -320,13 +320,13 @@ const About: React.FC = () => {
                         {team.map(({ name, avatar, role1, role2 }) => (
                             <div
                                 key={name}
-                                className="flex justify-start flex-col text-center"
+                                className="flex justify-start flex-col text-center not-prose"
                             >
                                 <img
                                     srcSet={`${avatar} 1500w`}
                                     src={avatar}
                                     alt={name}
-                                    className="w-full not-prose m-0 mb-6"
+                                    className="w-full m-0 mb-6"
                                 />
                                 <span
                                     className={clsx(
@@ -359,37 +359,53 @@ const About: React.FC = () => {
                                 )}
                             </div>
                         ))}
-                        <div className="flex flex-col justify-between text-center lg:justify-start">
+                        <div
+                            className={clsx(
+                                "flex",
+                                "flex-col",
+                                "justify-between lg:justify-start",
+                                "text-center",
+                            )}
+                        >
                             <div className="w-full not-prose m-0">
                                 <JoinUsIcon
-                                    className="m-0 w-full lg:mb-6"
+                                    className={clsx(
+                                        "m-0 w-full lg:mb-6",
+                                        "lg:h-[240px]",
+                                        "md:h-[220px]",
+                                        "h-[180px]",
+                                    )}
                                     isDark={colorMode === "dark"}
                                 />
                             </div>
-                            <a
-                                target="_blank"
-                                href="https://www.linkedin.com/company/refine-dev"
-                                className={clsx(
-                                    "text-xs leading-4",
-                                    "lg:text-base lg:leading-6",
-                                    "no-underline hover:no-underline text-refine-link-light dark:text-refine-link-dark font-semibold mb-0",
-                                )}
-                                rel="noreferrer"
-                            >
-                                Join our team!
-                            </a>
-                            <a
-                                target="_blank"
-                                href="https://www.linkedin.com/company/refine-dev"
-                                className={clsx(
-                                    "text-xs leading-4",
-                                    "lg:text-base lg:leading-6",
-                                    "no-underline hover:no-underline text-refine-link-light dark:text-refine-link-dark m-0",
-                                )}
-                                rel="noreferrer"
-                            >
-                                See open positions
-                            </a>
+                            <div className="flex flex-col items-center justify-center">
+                                <a
+                                    target="_blank"
+                                    href="https://www.linkedin.com/company/refine-dev"
+                                    className={clsx(
+                                        "block",
+                                        "text-xs leading-4",
+                                        "lg:text-base lg:leading-6",
+                                        "no-underline hover:no-underline text-refine-link-light dark:text-refine-link-dark font-semibold mb-0",
+                                    )}
+                                    rel="noreferrer"
+                                >
+                                    Join our team!
+                                </a>
+                                <a
+                                    target="_blank"
+                                    href="https://www.linkedin.com/company/refine-dev"
+                                    className={clsx(
+                                        "block",
+                                        "text-xs leading-4",
+                                        "lg:text-base lg:leading-6",
+                                        "no-underline hover:no-underline text-refine-link-light dark:text-refine-link-dark m-0",
+                                    )}
+                                    rel="noreferrer"
+                                >
+                                    See open positions
+                                </a>
+                            </div>
                         </div>
                     </div>
                 </div>