+++ examples/pixels-admin/vite.config.ts
@@ -15,10 +15,14 @@ export default defineConfig({
             },
             // Enable esbuild polyfill plugins
             plugins: [
+                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+                // @ts-ignore
                 NodeGlobalsPolyfillPlugin({
                     buffer: true,
                     process: true,
                 }),
+                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+                // @ts-ignore
                 NodeModulesPolyfillPlugin(),
             ],
         },