+++ packages/core/src/contexts/refine/IRefineContext.ts
@@ -32,6 +32,16 @@ export interface IRefineOptions {
     };
     overtime?: UseLoadingOvertimeRefineContext;
     textTransformers?: TextTransformers;
+    /**
+     * Disables server-side validation globally for the useForm hook
+     * @default false
+     * @see {@link https://refine.dev/docs/advanced-tutorials/forms/server-side-form-validation/}
+     */
+    disableServerSideValidation?: boolean;
+    /**
+     * The project id of your refine project. Will be set automatically. Don't modify.
+     */
+    projectId?: string;
 }
 
 export interface IRefineContextOptions {
@@ -48,6 +58,8 @@ export interface IRefineContextOptions {
     };
     overtime: UseLoadingOvertimeRefineContext;
     textTransformers: Required<TextTransformers>;
+    disableServerSideValidation: boolean;
+    projectId?: string;
 }
 
 export interface IRefineContext {