+++ packages/core/src/definitions/helpers/handleRefineOptions/index.ts
@@ -86,6 +86,10 @@ export const handleRefineOptions = ({
                 options?.textTransformers?.singular ??
                 defaultRefineOptions.textTransformers.singular,
         },
+        disableServerSideValidation:
+            options?.disableServerSideValidation ??
+            defaultRefineOptions.disableServerSideValidation,
+        projectId: options?.projectId,
     };
 
     const disableTelemetryWithDefault =