+++ packages/react-hook-form/test/index.tsx
@@ -1,13 +1,22 @@
 import React, { ReactNode } from "react";
 import { MemoryRouter } from "react-router-dom";
-import { Refine, DataProvider, IResourceItem } from "@refinedev/core";
+import {
+    Refine,
+    DataProvider,
+    IResourceItem,
+    I18nProvider,
+} from "@refinedev/core";
 
-import { MockRouterProvider, MockJSONServer } from "./dataMocks";
+import { MockJSONServer, mockRouterBindings } from "./dataMocks";
+import "@testing-library/jest-dom/extend-expect";
+import { IRefineOptions } from "@refinedev/core/dist/interfaces";
 
 interface ITestWrapperProps {
     dataProvider?: DataProvider;
     resources?: IResourceItem[];
     routerInitialEntries?: string[];
+    i18nProvider?: I18nProvider;
+    options?: IRefineOptions;
 }
 
 export const TestWrapper: (
@@ -16,16 +25,20 @@ export const TestWrapper: (
     dataProvider,
     resources,
     routerInitialEntries,
+    i18nProvider,
+    options,
 }) => {
     // eslint-disable-next-line react/display-name
     return ({ children }): React.ReactElement => {
         return (
             <MemoryRouter initialEntries={routerInitialEntries}>
                 <Refine
+                    i18nProvider={i18nProvider}
                     dataProvider={dataProvider ?? MockJSONServer}
-                    routerProvider={MockRouterProvider}
+                    routerProvider={mockRouterBindings()}
                     resources={resources ?? [{ name: "posts" }]}
                     options={{
+                        ...options,
                         reactQuery: {
                             clientConfig: {
                                 defaultOptions: {